Webpack common dist config now supports much lighter dist configuration files, eliminating any duplicate build configuration code. Config merging occurs as such: webpackCommonCfg -> webpackDistCfg -> masterDistCfgFile. Dev env configs are unaffected (though the dev config was modified to be simpler). Added favicon to the example project but need to add settings to be able to target it specifically.
